Change int16 to int to be compatible with asn1.
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@1119 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/krb5/krb5.h
@@ -79,7 +79,7 @@ enum {
 };
 
 typedef struct krb5_address{
-    int16_t type;
+    int type;
     krb5_data address;
 } krb5_address;
